work on cleanup, work on snapshotting light logic
