+++ src/main/java/net/minestom/server/instance/light/Neighbors.java
@@ -1,0 +1,40 @@
+package net.minestom.server.instance.light;
+
+public enum Neighbors {
+    NORTH(0, -1),
+    EAST(1, 0),
+    SOUTH(0, 1),
+    WEST(-1, 0),
+    NORTH_EAST(1, -1),
+    NORTH_WEST(-1, -1),
+    SOUTH_EAST(1, 1),
+    SOUTH_WEST(-1, 1);
+
+    private final int x, z;
+
+    Neighbors(int x, int z) {
+        this.x = x;
+        this.z = z;
+    }
+
+    public int x() {
+        return x;
+    }
+
+    public int z() {
+        return z;
+    }
+
+    public Neighbors opposite() {
+        return switch (this) {
+            case NORTH -> SOUTH;
+            case EAST -> WEST;
+            case SOUTH -> NORTH;
+            case WEST -> EAST;
+            case NORTH_EAST -> SOUTH_WEST;
+            case NORTH_WEST -> SOUTH_EAST;
+            case SOUTH_EAST -> NORTH_WEST;
+            case SOUTH_WEST -> NORTH_EAST;
+        };
+    }
+}
