+++ src/main/java/net/minestom/server/instance/light/parallel/package-info.java
@@ -1,0 +1,4 @@
+@NotNullByDefault
+package net.minestom.server.instance.light.parallel;
+
+import org.jetbrains.annotations.NotNullByDefault;
