+++ src/main/java/net/minestom/server/instance/light/snapshot/OcclusionData.java
@@ -1,0 +1,4 @@
+package net.minestom.server.instance.light.snapshot;
+
+public record OcclusionData(int highestBlock, int[] occlusionMap) {
+}
