+++ src/main/java/net/minestom/server/instance/light/snapshot/package-info.java
@@ -1,0 +1,4 @@
+@NotNullByDefault
+package net.minestom.server.instance.light.snapshot;
+
+import org.jetbrains.annotations.NotNullByDefault;
